separate out parsing_xml
